requirements.txt
agno>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
google-generativeai>=0.3.0
sqlalchemy>=2.0.0
uvicorn>=0.20.0
fastapi>=0.100.0
streamlit
google-genai